feat(markdown_viewer): making the markdown parser less noisy.
from now alot of the elements will just be highlighted instead of having
custom ANSI format -- most people are familiar with markdown formatting
anyways, and its readable as is.

also from now HTML elements will be commentted. they mostly just for
layouting, and the previous highlighting approach made it impossible to
read the markdown when HTML is present.